moving delete channel intent handlers to separate folder
